scraped P7 in favor of hotfixes

+++ CSCI-4526-Sudoku/P7-KorideMok/DiagBoard-KorideMok.hpp
@@ -8,7 +8,7 @@
 class DiagBoard : public Board {
     public:
         DiagBoard() = default;
-        DiagBoard(int, int, ifstream&);
+        DiagBoard(short, short, ifstream&);
         ~DiagBoard() = default;
         void createDiagonal();
 };
